split config into 2 files
