+++ webapp/src/app/api/messages/[projectName]/route.ts
@@ -1,5 +1,6 @@
+import { LyraConfig } from '@/utils/lyraConfig';
 import MessageAdapterFactory from '@/utils/adapters/MessageAdapterFactory';
-import { LyraConfig, ServerConfig } from '@/utils/lyraConfig';
+import { ServerConfig } from '@/utils/serverConfig';
 import {
   LyraConfigReadingError,
   ProjectNameNotFoundError,
